Create scroll_down function to address loading issue
Using a simple get_element command is not sufficient to obtain
all available houses matching search criteria on Seabrook
because the listings load as you scroll down.

scroll_down addresses this by scrolling up and down on the page.
Note that for this to work, the browser window must be maximized.