Add looping over search results
Create a loop to open each search result link in the same window.
Also re-name a couple of files to better reflect the architecture
of the project.

navtools contains methods called by main to get
information from Seabrook's website. Note that this approach does
not allow tracking of window handles. This should not be a problem
as once the necessary information has been obtained from each page,
it can be closed and will not need to be revisited.